Update device restart - clarification
changed the first sentence from:


The **Restart** device action causes the device you choose to be restarted. The device owner isn't automatically notified of the restart, and they might lose work.
 
to

The **Restart** device action causes the device you choose to be restarted. The end user isn't automatically notified of the restart, and they might lose work.

Because "Device Owner" needs to be designated as the supported device type in Android, this will cause confusion

Changed Android kiosk devices - Supported
to
Android Device Owner - Supported

Kiosk is not the correct terminology and doesn't consider the other Android Device owner scenarios (e.g. COSU & COBO, potentially COPE in the future)